[FIX] Removed unused import in app.py, refactored test_app.py
